Change meme dog file structure to match the 2 newer fights
