+++ lib/meme_dog/__init__.py
@@ -1,0 +1,1 @@
+from .meme_dog import MemeDog
